+++ component.mk
@@ -1,3 +1,3 @@
 COMPONENT_SRCDIRS := src
-COMPONENT_INCDIRS := src
-COMPONENT_DOXYGEN_INPUT := src
+COMPONENT_INCDIRS := src/include
+COMPONENT_DOXYGEN_INPUT := src/include
