+++ src/include/MHZ19.h
@@ -1,0 +1,23 @@
+/**
+ * MHZ19.cpp
+ *
+ * Copyright 2021 mikee47 <mike@sillyhouse.net>
+ *
+ * This file is part of the Sming-MHZ19 Library
+ *
+ * This library is free software: you can redistribute it and/or modify it under the terms of the
+ * GNU General Public License as published by the Free Software Foundation, version 3 or later.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with this library.
+ * If not, see <https://www.gnu.org/licenses/>.
+ *
+ ****/
+
+#pragma once
+
+#include "MHZ19/Uart.h"
+#include "MHZ19/PwmReader.h"
