+++ src/include/MHZ19/Uart.h
@@ -21,6 +21,7 @@
 
 #include <HardwareSerial.h>
 #include <SimpleTimer.h>
+#include "common.h"
 
 namespace MHZ19
 {
@@ -57,12 +58,6 @@ enum class Error {
 	incompleteResponse,
 	invalidResponse,
 	timeout,
-};
-
-enum class DetectionRange {
-	PPM_2000 = 2000,
-	PPM_5000 = 5000,
-	PPM_10000 = 10000,
 };
 
 struct Measurement {
@@ -151,54 +146,6 @@ private:
 	SimpleTimer timer;
 	MeasurementCallback callback;
 };
-
-/**
- * @brief Reads input pulse width asynchronously
- */
-class PwmReader
-{
-public:
-	union Pulse {
-		struct {
-			uint16_t low;
-			uint16_t high;
-		};
-		uint32_t value;
-	};
-
-	~PwmReader()
-	{
-		end();
-	}
-
-	void begin(uint8_t pin, DetectionRange range = DetectionRange::PPM_2000);
-
-	void end();
-
-	uint16_t getMeasurement() const;
-
-private:
-	static void IRAM_ATTR staticInterruptHandler()
-	{
-		return self->interruptHandler();
-	}
-
-	void interruptHandler();
-
-	static PwmReader* self;
-	uint8_t pin;
-	uint32_t isrTicks{0};
-	Pulse isrPulse{};
-	volatile Pulse reading{};
-	DetectionRange range;
-};
-
-/**
- * @brief Read PWM output from sensor
- * @param pwmPin GPIO to which the sensor is connected
- * @param range Range sensor is configured for
- */
-unsigned pwmRead(uint8_t pwmPin, DetectionRange range = DetectionRange::PPM_2000);
 
 } // namespace MHZ19
